+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 The complete changelog for the Costs to Expect REST API, our changelog follows the format defined at https://keepachangelog.com/en/1.0.0/
 
 ## [0.5.0] - [Date]
+### Added
+- Added a complete game action, sets the winner, player scores and updates the game status.
+
 ### Changed
 - Show the player scores next to open games.